Updated from upstream.
