+++ .gitignore
@@ -1,1 +1,2 @@
 .Rproj.user
+src/*.o
