+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: arulesCBA
-Version: 1.0.3
-Date: 2016-08-28
+Version: 1.1.0
+Date: 2017-02-15
 Title: Classification Based on Association Rules
-Author: Ian Johnson, Michael Hahsler
+Author: Ian Johnson [aut, cre, cph], Michael Hahsler [ctb]
 Authors@R: c(person("Ian", "Johnson", role=c("aut", "cre","cph"),
             email = "ianjjohnson@icloud.com"),
             person("Michael", "Hahsler", role=c("ctb")))
