+++ README.md
@@ -4,8 +4,13 @@
 [![CRAN RStudio mirror downloads](http://cranlogs.r-pkg.org/badges/arulesCBA)](https://cran.r-project.org/package=arulesCBA)
 [![CRAN version](http://www.r-pkg.org/badges/version/arulesCBA)](https://cran.r-project.org/package=arulesCBA)
 
-This R package implements the CBA algorithm described in Liu, et al. 1998.
-It creates classifiers based on association rules and can then use those classifiers to classify incoming datasets.
+This R package performs association rule-based classification.
+
+It includes two classification algorithms, both of which are bundled in the CBA function.
+The first is the CBA algorithm described in Liu, et al. 1998.
+The second is a weighted majority-vote based algorithm which is currently being designed and tested.
+
+CBA creates classifiers based on association rules and can then use those classifiers to classify incoming datasets.
 
 The algorithms are implemented in C, and data is formatted and passed to the C implementations via an R interface accessible through this package.
